refactor: Splitter 클래스에 피연산자를 추출하는 extractOperands 메서드 작성
